Fix codeclimate error
